2D/Canvas: Text2D/Label::set_font_size() bug. Object: Invalid expression on __GetPropertyType. Physics: Changing the use of BodyType, now KINEMATIC has zero mass. Platform/Android: Implementation of File::exists(), fixing format string in main, now AINPUT_EVENT_TYPE_KEY is handled correctly.

+++ src/scene/scene_manager.cpp
@@ -266,7 +266,7 @@ void SceneManager::scene_handle_input(const InputEvent& event)
 
 void SceneManager::_try_clear_root_canvas()
 {
-    for (i32 i = 0; i < data.root_canvas.count; i++)
+    for (usize i = 0; i < data.root_canvas.count; i++)
     {
         CanvasObject* gui_root = data.root_canvas[i];
